Support StatsModels v0.6 with `FormulaTerm`
This is a different approach to upgrading support for StatsModels to 0.6. It
takes advantage of Terms 2.0 (Son of Terms) functionality by creating a special
Term type that wraps the left-hand side of the formula, and returns a sampled
version of the response variable every time it's asked to generate model
columns. This allows the two separate methods for bootstrap using different
sampling strategies to be combined by factoring out the response re-sampling
into a single place. More importantly, making the resampling step lazy in this
way removes all of the bookkeeping and copying of data, since the residuals are
resampled and added to the response at the last moment.

It also allows for any of the terms types supported by terms 2.0 to be used on
the LHS, including arbitrary functions. So you can do things like
@formula(log10(X) ~ log10(U)) and it will work (and the tests have been updated
to reflect this)

Includes:
* bump statsmodels to 0.6 and Formula->FormulaTerm
* guard against FunctionTerm on LHS of formula
* use ResidualTerm for resampling
* use transformed formulae in parameteric bootstrap tests